backend/requirements.txt
fastapi
uvicorn
python-dotenv
pinecone
google-genai